Add app version via mvn
